Rename extrinsics weights module
