+++ pallets/rad-claims/src/weights.rs
@@ -15,11 +15,35 @@
 // along with Cumulus.  If not, see <http://www.gnu.org/licenses/>.
 
 
-//! Rad claims pallet's extrinsics weight information
+//! Crowdloan claim pallet's extrinsics weight information
 //! 
-//! This module must contain the weights that are calculated using
-//! runtime benchmarking. 
-//! This file will be replaced when you first run benchmarking.
-//! You should first write various benchmarking cases in the 
-//! [`benchmarking`](./benchmarking.rs) module.
+//! Note that the following weights are used only for development.
+//! In fact, weights are calculated using runtime benchmarking.
 
+use frame_support::{
+    weights::{
+        constants::RocksDbWeight,
+        Weight,
+    }
+};
+
+use crate::traits::WeightInfo;
+
+
+impl WeightInfo for () {
+
+    fn claim( hashes_length: usize ) -> Weight {
+        (195_000_000 as Weight).saturating_add(
+            hashes_length.saturating_mul(1_000_000) as Weight +
+            RocksDbWeight::get().reads_writes(2,2)
+        )
+    }
+
+    fn set_upload_account() -> Weight { 
+        190_000_000 as Weight 
+    }
+
+    fn store_root_hash() -> Weight { 
+        185_000_000 as Weight 
+    }
+}
